requirements.txt
PyYAML==6.0.2
requests==2.32.3
imap-tools==1.8.0
